+++ src/contracts/interfaces/IFNFTFactory.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.0;
 import {IPriceOracle} from "./IPriceOracle.sol";
 import {IWETH} from "./IWETH.sol";
 
-interface IFNFTFactory {    
+interface IFNFTFactory {
     function vaultManager() external view returns (address);
 
-    function swapFee() external view returns (uint256);
+    function swapFee() external view returns (uint64);
 
     function maxAuctionLength() external view returns (uint256);
